Fixed images on carousel not updating
The problem was that the carousel doesn't work well with next/image. next/image isn't even that needed, since the images are hosted on Supabase CDN after all. The only remaining problem is that when we loop the carousel, it takes about 0.5ms for it to loop back. I fixed this by removing looping on the carousel, but it would be nice to find a solution for this in the future.